Combed through all files and made sure they are all named semantically and consistently. Organized the molecules directory.
