+++ src/components/Molecules/ConnectMolecules/connectInfo.css
