+++ src/components/Molecules/HomeMolecules/BookNow/bookNow.js
@@ -1,6 +1,6 @@
 import "./bookNow.css";
-import Logo from "../../Atoms/Pictures/Logo/Logo";
-import Button from "../../Atoms/Buttons/HeaderButton/HeaderButton";
+import Logo from "../../../Atoms/Pictures/Logo/Logo";
+import Button from "../../../Atoms/Buttons/HeaderButton/HeaderButton";
 
 export default function bookNow() {
   return (
