+++ src/components/Molecules/HomeMolecules/ContactSection/ContactSection.js
@@ -1,4 +1,4 @@
-import Form from "../../Form/Form";
+import Form from "../Form/Form";
 import "./contactSection.css";
 
 export default function ContactSection() {
