+++ src/components/Molecules/HomeMolecules/ServicesSection/ServicesSection.js
@@ -1,6 +1,6 @@
 import "./servicesSection.css";
 import ServicesText from "../../../Atoms/Text/HomePage/ServicesText/HeaderText";
-import ServicesBox from "../../../Molecules/ServicesBox/ServicesBox";
+import ServicesBox from "../ServicesBox/ServicesBox";
 import ServicesButton from "../../../Atoms/Buttons/ServicesButton/ServicesButton";
 
 export default function Services() {
