+++ src/components/Molecules/HomeMolecules/TestimonialSection/TestimonialsSection.js
@@ -1,5 +1,5 @@
 import "./testimonialsSection.css";
-import TestimonialsBox from "../../../Molecules/TestimonialsBox/TestimonialsBox";
+import TestimonialsBox from "../TestimonialsBox/TestimonialsBox";
 import HomeWhy from "../../../Atoms/Text/HomePage/TestimonialText/TestimonialText";
 
 export default function Testimonials() {
