+++ src/components/Molecules/ReusableMolecules/Footer/Footer.js
@@ -1,6 +1,6 @@
 import "./footer.css";
 import Socials from "../SocialLinks/SocialLinks";
-import LogoFooter from "../../Atoms/Pictures/Logo/LogoFooter";
+import LogoFooter from "../../../Atoms/Pictures/Logo/LogoFooter";
 
 export default function Footer() {
   return (
