+++ src/components/Organisms/AboutPage/AboutContainer.js
@@ -1,7 +1,7 @@
 import "./aboutContainer.css";
-import NavBar from "../../Molecules/NavBar/NavBar";
-import Footer from "../../Molecules/Footer/Footer";
-import AboutSection from "../../Molecules/AboutSection/AboutSection";
+import NavBar from "../../Molecules/ReusableMolecules/NavBar/NavBar";
+import Footer from "../../Molecules/ReusableMolecules/Footer/Footer";
+import AboutSection from "../../Molecules/AboutMolecules/AboutSection";
 
 export default function AboutContainer() {
     return (
