--- src/components/Organisms/ConnectPage/Connect.js
@@ -1,14 +1,0 @@
-import "./connect.css";
-import NavBar from "../../Molecules/NavBar/NavBar";
-import Footer from "../../Molecules/Footer/Footer";
-import PageContainer from "../../Molecules/ConnectSection/PageContainer"
-
-export default function ConnectContainer() {
-  return (
-    <div className="connectPageContainer">
-      <NavBar />
-      <PageContainer />
-      <Footer />
-    </div>
-  )
-}
