+++ src/components/Organisms/HomePage/HomepageContainer.js
@@ -3,8 +3,8 @@ import HomeHeader from "../../Molecules/HomeMolecules/Header/Header";
 import Testimonials from "../../Molecules/HomeMolecules/TestimonialSection/TestimonialsSection";
 import Services from "../../Molecules/HomeMolecules/ServicesSection/ServicesSection";
 import Contact from "../../Molecules/HomeMolecules/ContactSection/ContactSection";
-import NavBar from "../../Molecules/NavBar/NavBar";
-import Footer from "../..//Molecules/Footer/Footer";
+import NavBar from "../../Molecules/ReusableMolecules/NavBar/NavBar";
+import Footer from "../../Molecules/ReusableMolecules/Footer/Footer";
 
 export default function HomepageContainer() {
     return (
