+++ src/pages/ServicesPage.js
@@ -1,4 +1,4 @@
-import ServicesContainer from "../components/Organisms/ServicesPage/ServicesPage"
+import ServicesContainer from "../components/Organisms/ServicesPage/ServicesContainer"
 
 export default function Services() {
   return (
